requirements.txt
asgiref==3.6.0
boto3==1.26.137
botocore==1.29.137
certifi==2023.5.7
charset-normalizer==3.1.0
colorama==0.4.6
decorator==4.4.2
Django==4.1.9
django-cors-headers==4.0.0
djangorestframework==3.14.0
djongo==1.3.6
dnspython==2.3.0
idna==3.4
imageio==2.28.1
imageio-ffmpeg==0.4.8
jmespath==1.0.1
moviepy==1.0.3
numpy==1.24.3
pika==1.3.2
Pillow==9.5.0
proglog==0.1.10
pymongo==3.12.3
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
requests==2.30.0
s3transfer==0.6.1
six==1.16.0
sqlparse==0.2.4
tqdm==4.65.0
tzdata==2023.3
urllib3==1.26.15
